move startupcheck to commons
